# docu-rag/api/openapi.yaml
openapi: 3.0.3
info:
  title: DocuRAG REST API
  description: |
    Medical document RAG system — Spring Boot Modulith.
    
    ## Architecture Overview
    Documents are ingested (PDF or structured JSONL/CSV), stored as source documents,
    chunked into passages, and embedded into a pgvector store for semantic search.
    
    ## Typical Flow
    1. **Ingest** documents via `POST /api/documents/ingest`
    2. **Build index** via `POST /api/index/rebuild`
    3. **Query** via `POST /api/rag/ask`
    
    E2E client source of truth — keep aligned with controllers under `com.berdachuk.docurag.web.rest`.
  version: 0.1.0
servers:
  - url: http://localhost:8084
    description: Local default
  - url: http://localhost:18080
    description: Local E2E / stub-AI profile
...
paths:
  /api/documents/ingest:
    post:
      tags: [documents]
      summary: Ingest documents from specified paths
      description: |
        Ingests documents from file paths or directories. Supports PDF files and
        structured formats (JSONL, CSV). Documents are parsed, deduplicated by content hash,
        and stored. For PDF files, text is extracted directly. For structured files, each
        record is treated as a separate document. After ingestion, use /api/index/rebuild
        to create chunks and embeddings.
      operationId: ingestDocuments
      requestBody:
        description: Paths to files or directories to ingest
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestPathsRequest'
      responses:
        '200':
          description: Ingest completed — returns job summary with loaded/skipped counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestSummary'
  /api/documents:
    get:
      tags: [documents]
      summary: List ingested documents
      description: Returns a paginated list of all ingested source documents with their metadata.
      operationId: listDocuments
      parameters:
        - name: page
          in: query
          description: Page number (zero-based)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of document summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceDocumentSummary'
...
  /api/evaluation/run:
    post:
      tags: [evaluation]
      summary: Run evaluation on a dataset
      description: |
        Runs evaluation against a built-in or custom dataset. Measures RAG quality
        using ground-truth questions and expected answers. Returns aggregate metrics.
      operationId: runEvaluation
      requestBody:
        description: Evaluation parameters and dataset
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRunRequest'
      responses:
        '200':
          description: Evaluation run started and completed synchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationRunStartedResponse'
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Cannot run in current system state
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  schemas:
    IngestPathsRequest:
      type: object
      properties:
        paths:
          type: array
          items:
            type: string
    IngestSummary:
      type: object
      properties:
        jobId:
          type: string
        documentsLoaded:
          type: integer
          format: int32
        documentsSkipped:
          type: integer
          format: int32
        status:
          type: string
        errorMessage:
          type: string
          nullable: true
    SourceDocumentSummary:
      type: object
      properties:
        id:
          type: string
        externalId:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        sourceFormat:
          type: string
          nullable: true
...
    EvaluationRunRequest:
      type: object
      required: [datasetName]
      properties:
        datasetName:
          type: string
        topK:
          type: integer
          format: int32
          minimum: 1
        minScore:
          type: number
          format: double
          minimum: 0
          maximum: 1
        semanticPassThreshold:
          type: number
          format: double
          minimum: 0
          maximum: 1
    EvaluationRunStartedResponse:
      type: object
      properties:
        runId:
          type: string
        total:
          type: integer
          format: int32
        normalizedAccuracy:
          type: number
          format: double
        meanSemanticSimilarity:
          type: number
          format: double
        semanticAccuracy:
          type: number
          format: double
        semanticPassThreshold:
          type: number
          format: double
        semanticAccuracyAt080:
          type: number
          format: double
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          format: uri
          nullable: true
      additionalProperties: true